refactor: signIn, signUpWithPhone del
