+++ lib/presentation/providers/datepicker_provider.dart
@@ -13,12 +13,11 @@ class DatePickerProvider with ChangeNotifier {
         context: context, firstDate: DateTime(1900), lastDate: DateTime.now());
 
     if (picked != null && picked != _selectedDate) {
-      {
         _selectedDate = picked;
         _dateController.text =
             DateFormat('yyyy/MM/dd').format(picked); // 선택된 날짜를 텍스트 필드에 표시
-      }
-    }
+
     notifyListeners();
+    }
   }
 }
