+++ lib/presentation/views/home_screen.dart
@@ -4,8 +4,9 @@ import 'package:metabugo/presentation/views/account/account_add_screen.dart';
 import 'package:metabugo/presentation/views/funeral/funeral_screen.dart';
 import 'package:metabugo/presentation/views/funeral_send_list_screen.dart';
 import 'package:metabugo/presentation/views/privacy_agreement_modal.dart';
+import 'package:metabugo/presentation/views/send_flower.dart';
+import 'package:metabugo/presentation/views/signUp/sign_in.dart';
 import 'package:provider/provider.dart';
-import 'package:metabugo/presentation/views/sign_in.dart';
 import 'package:metabugo/res/media_res.dart';
 
 class HomeScreen extends StatelessWidget {
@@ -149,7 +150,7 @@ class _HomeScreenContent extends StatelessWidget {
                       .updateTextUnderLineColor3(MediaRes.selectColor);
                   Navigator.push(
                     context,
-                    MaterialPageRoute(builder: (context) => AccountAddScreen()),
+                    MaterialPageRoute(builder: (context) => SendFlower()),
                   ).then((_) {
                     homeScreenProvider.resetColors();
                   });
